Hi @krausest,

This is a from-scratch implementation of an interactive results UI. This work builds on top of the necessary minor tweaks in #418.

The original motivation for this was to reduce `node_modules` bloat in `webdriver-ts-results` (~100MB, ~13,000 files) and perhaps improve build times, bundle size as well as boost UI performance. I'm pleased to report that the final result has surpassed my original expectations by a wide margin.

Full feature parity with `webdriver-ts-results` was not a goal, so some things got the axe:

- **Use median instead of mean** - recent improvements in the benchmark runner have made the differences mostly irrelevant
- **Separate keyed and non-keyed** - the ability to display keyed and non-keyed frameworks in the same table is gone
- **Compare with...** - i never really understood how this feature was supposed to work, including why it has a drastically different color scale. How is it different than simply enabling only the frameworks you're interested in comparing?
- **building results from `src/results.ts`** - the results-ui now only builds from the raw `.json` files produced by `webdriver-ts` because it doesn't matter what that file contains since all its contents are from a different environment. if you don't have the original `.json` files produced on your own machine, they may as well not exist.
- **CSS** - dropped bootstrap due to huge size in favor of custom css

The important stuff is still there:

- Toggle frameworks
- Toggle benchmarks
- Sort by benchmark, name or total slowdown
- Reverse sorting **NEW**

Additional notes:

- Some bench names and descriptions no longer represented what they actually did, so I adjusted the names to be more-or-less uniform and accurate.
- The color scale is not identical but it's quite close. Rather than using a smooth scaling function, I generated a 16 color scale [1] that closely resembles the current one and selected the step thresholds [2] carefully to match what exists now. This has a good UI performance benefit since classes are much more likely to be shared during mutation and do not need to change as often as style attributes tweaked precisely to the factor.
- For better compression, I reduced the precision of all measurements to 2 decimal places, followed by encoding all benchmarks into arrays rather than objects. These are inflated at run-time with minimal overhead, but reduce the bundle size enormously.
- The implementation is based on domvm, rollup/buble and uglify-js
- There are no vdom optimizations in this code - the whole vdom is regenerated and diffed for all interactions. This is done for simplicity, but there's still some perf left on the table, if needed.

Now for some fun stats and pictures:

- **node_modules:** 97.6MB (12,927 files) -> 12.0MB (536 files)
![dependencies](https://user-images.githubusercontent.com/43234/43110921-2fe1e928-8eb4-11e8-8cc6-bbd4ce21bba5.png)
- **build time & `table.html` size:** 18.3s -> 2.7s, 1MB -> 197KB
![build-times](https://user-images.githubusercontent.com/43234/43110920-2fd253fa-8eb4-11e8-924b-004ec31e0297.png)
- **initial render (scripting):** 440ms -> 100ms
![load](https://user-images.githubusercontent.com/43234/43110922-2ff7796e-8eb4-11e8-8ffc-788d61afe825.png)
- **sorting by name, each benchmark and total (scripting):** 2333ms -> 260ms
![sorting-11x](https://user-images.githubusercontent.com/43234/43110923-3009cc54-8eb4-11e8-8ace-10ca86fd6caa.png)
- **toggling each keyed bench 2x (scripting):** 2090ms -> 450ms
![toggle-benches](https://user-images.githubusercontent.com/43234/43110924-3018e554-8eb4-11e8-95e0-b3f56273a6e4.png)

(The build time improvement is actually even more drastic since `webdriver-ts-results` was building from a prepared `results.ts` while `results-ui` was synchronously reading/parsing 1,656 `.json` files from `webdriver-ts/results`)

[1] https://www.strangeplanet.fr/work/gradient-generator/?c=16:63BF7C:FFEB84:F9696C
[2] https://github.com/leeoniya/js-framework-benchmark/blob/results-ui/results-ui/src/ui.js#L102-L107