@krausest 2018-07-24
Thanks. I'll take a look at it when I rerun the results with Chrome 68.

I'm not completely sure that we're after the same goals though. I don't care too much about dependency size of the result-ui (I have 94 implementations and some with large dependencies - I'm not sure if this is the place to start sizing down) though faster build times are very welcome of course.

The compare with feature is essential to me. It's such a shame that I failed to explain what it does. Put shortly: If you compare against a framework and you don't get either a red or green box for the other framework you still might say one is faster than the other, but you can't back your claim on statistical significance.
Benchmarks neglecting statistic tests can be easily criticized and I'm working hard not to be.

I think at first I'll add your results-ui as an alternate results-renderer and gradually merge it, but I want to keep (p)react [sorry!], compare with, gradient results and mean/median. I'll also merge #418 for the new results.

Maybe we should start a new benchmark from the results renderer? By now we have a react, preact, vue and domvm implementation  :smirk: