@leeoniya 2018-07-24
i probably have an irrational fear of huge dependencies. part of it might be due to the fact that i run many things portably (off a USB flash drive) so installing/deleting/updating 13,000 files is significantly worse than 500 files.

> Put shortly: If you compare against a framework and you don't get either a red or green box for the other framework you still might say one is faster than the other, but you can't back your claim on statistical significance.

do you have any examples from the current data where the mean and/or median are misleading, but get saved by sound statistics (besides < 20ms tests, where variance is wild).

> I'll also merge #418 for the new results.

that would be great.

thanks!