@krausest 2018-07-26
I had a small problem with missing results and tried to fix it  - might not be the most elegant way.