@leeoniya 2018-07-26
fix looks good, though i try to avoid truthy/falsey coercions in property checks, preferring `x.y == null` to `!x.y`. also, if you use spaces, it probably makes sense to convert the whole file to spaces. your changes makde it mixed spaces/tabs.