@krausest 2018-08-06
@leeoniya Thanks for the review, it was badly needed. I reviewed the code and compared it to the results of the nice web site above and excel and found 3 (!) mistakes:
* std deviation was computed incorrectly: missing true parameter to `jStat(result.values).stdev(true)`
* Sum of squares was incorrect bracketed (oh no...)
* Colouring started actually with a  p-Value of 20 instead of 10
I'm now printing the p-Value below the factor such that checking the values should be easier.
<img width="284" alt="bildschirmfoto 2018-08-06 um 21 37 21" src="https://user-images.githubusercontent.com/3374055/43737237-f04d4de4-99c0-11e8-9e9e-ffa041f5e77e.png">
The web site you found computes almost the same value:
<img width="832" alt="bildschirmfoto 2018-08-06 um 21 38 33" src="https://user-images.githubusercontent.com/3374055/43737338-3f85172a-99c1-11e8-9527-160cd2407021.png">
And now the result looks like that:
<img width="605" alt="bildschirmfoto 2018-08-06 um 21 52 24" src="https://user-images.githubusercontent.com/3374055/43737894-11f8b364-99c3-11e8-8274-86834883f160.png">
Looking forward your review!