@leeoniya 2018-08-06
@krausest 

DISCLAIMER: i'm not a statistician, so treat everything i say with skepticism.

a typical t-test is double-sided [1] in that its p value that represents the likelyhood of being wrong in both directions. (e.g.: mean is *either* greater *or* smaller).

maybe you want to take only the implied side of hypothesis, so in this case, `d ≥ 0` for inferno >=ivi. which gives you the `p` value for the mean only being greater rather than also including the chance that it was less. this halves the p value from 0.00466 to 0.00233. i'm not really sure if this is appropriate here, but since the colors imply only a specific side of the assumption, maybe it makes sense?

one thing i find confusing is why the p-value drops when you only take one side. it's not like excluding the probability of being wrong in the opposite direction makes you more certain of the other direction.

also, since we could be comparing metrics with differing variances, a Welch's t-test [2] would be more appropriate.

[1] https://stats.idre.ucla.edu/other/mult-pkg/faq/general/faq-what-are-the-differences-between-one-tailed-and-two-tailed-tests/
[2] https://en.wikipedia.org/wiki/Welch%27s_t-test