@leeoniya 2018-08-07
also, i'm pretty sure you can't do a t-test using medians: https://github.com/krausest/js-framework-benchmark/blob/master/webdriver-ts-results/src/Common.tsx#L248

if you plan to keep using jStat, i recommend switching over to using a Welch's t-test unconditionally [1] (steal code from here [2]). still not sure if a 2-tailed or 1-tailed test is appropriate here :/

[1] http://daniellakens.blogspot.com/2015/01/always-use-welchs-t-test-instead-of.html
[2] https://www.icalcu.com/stat/two-sample-t-test-calculator.html