@krausest 2018-08-07
Using medians makes indeeed no sense for the t-test.

I've now switched to welch's t-test. Excel and my program report the same values.
The resulting p-Values are very similar to a t-test (not sure how bad the assumption of equal variance was...)
Since I'm currently colouring p-Values below 10% (which is rather high) for a two-sided test it is  equivalent to one-sided tests at 5% (which is okay). So I see no reason to change that.