@leeoniya 2018-08-07
nice.

> The resulting p-Values are very similar to a t-test (not sure how bad the assumption of equal variance was...)

either way, Welch's properly handles different sample sizes as well as variances...all without any downsides. at the end of the day, reducing variance should be a primary goal of this bench, so hopefully we dont have to rely on being saved by Mr. Welch :)